external/ESM/requirements.txt
python-dotenv
esm
tqdm